refactor(filter-transactions): move to filter folder
