+++ src/features/filter/index.ts
@@ -1,0 +1,4 @@
+export interface TabItem<T = string> {
+  i18nKey: string
+  value: T
+}
