+++ src/pages/AccountDetails.vue
@@ -10,8 +10,8 @@ import type { AssetId } from '@iroha/core/data-model';
 import { AccountIdSchema } from '@/shared/api/schemas';
 import { parseMetadata } from '@/shared/ui/utils/json';
 import BaseTabs from '@/shared/ui/components/BaseTabs.vue';
-import type { TabAccountTransactions } from '@/features/filter-transactions/model';
-import { ACCOUNT_TRANSACTIONS_OPTIONS } from '@/features/filter-transactions/model';
+import type { TabAccountTransactions } from '@/features/filter/transactions/model';
+import { ACCOUNT_TRANSACTIONS_OPTIONS } from '@/features/filter/transactions/model';
 import TransactionsTable from '@/shared/ui/components/TransactionsTable.vue';
 import InstructionsTable from '@/shared/ui/components/InstructionsTable.vue';
 import BaseLink from '@/shared/ui/components/BaseLink.vue';
