+++ src/shared/ui/components/BaseTabs.vue
@@ -47,7 +47,7 @@ import type { AdaptiveOptions } from '@/shared/ui/utils/adaptive-options';
 import { applyAdaptiveOptions } from '@/shared/ui/utils/adaptive-options';
 import ArrowIcon from '@soramitsu-ui/icons/icomoon/arrows-chevron-left-rounded-24.svg';
 import { useI18n } from 'vue-i18n';
-import type { TabItem } from '@/features/filter-transactions/model';
+import type { TabItem } from '@/features/filter';
 
 const { t } = useI18n();
